+++ packages/style/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.1](https://github.com/mondaycom/vibe/compare/monday-ui-style@0.14.0...monday-ui-style@0.14.1) (2024-09-29)
+
+**Note:** Version bump only for package monday-ui-style
+
+
+
+
+
 # [0.14.0](https://github.com/mondaycom/vibe/compare/monday-ui-style@0.13.0...monday-ui-style@0.14.0) (2024-09-26)